Add templates with view layout and tests
