--- lib/inquisition/outputter/doc/layout.rb
@@ -1,19 +1,0 @@
-require_relative 'home'
-
-module Inquisition
-  module Outputter
-    class Doc
-      class Layout
-        def produce
-          binding
-        end
-
-        def home
-          @home ||= begin
-            Template.new('home').render(Home.new)
-          end
-        end
-      end
-    end
-  end
-end
